+++ FirebaseMessaging/Sources/Public/FirebaseMessaging/FIRMessaging.h
@@ -65,7 +65,8 @@ typedef void (^FIRMessagingTopicOperationCompletion)(NSError *_Nullable error)
 // clang-format off
 // clang-format12 merges the next two lines.
 FOUNDATION_EXPORT const NSNotificationName FIRMessagingRegistrationTokenRefreshedNotification
-    NS_SWIFT_NAME(MessagingRegistrationTokenRefreshed);
+    NS_SWIFT_NAME(MessagingRegistrationTokenRefreshed)
+        DEPRECATED_MSG_ATTRIBUTE("Use messaging:didReceiveRegistration: instead.");
 
 /**
  *  Notification sent when the FCM installation id has been unregistered.
@@ -160,7 +161,8 @@ NS_SWIFT_NAME(MessagingDelegate)
 /// * Subscribing to any topics.
 - (void)messaging:(FIRMessaging *)messaging
     didReceiveRegistrationToken:(nullable NSString *)fcmToken
-    NS_SWIFT_NAME(messaging(_:didReceiveRegistrationToken:));
+    NS_SWIFT_NAME(messaging(_:didReceiveRegistrationToken:))
+        DEPRECATED_MSG_ATTRIBUTE("Use messaging:didReceiveRegistration: instead.");
 
 /// This method will be called once the registration is created or refreshed. When
 /// auto init is enabled, it will be automatically called once per app start, but may be called
@@ -307,7 +309,8 @@ NS_SWIFT_NAME(Messaging)
  * Once you have an FCM registration token, you should send it to your application server, where
  * it can be used to send notifications to your device.
  */
-@property(nonatomic, readonly, nullable) NSString *FCMToken NS_SWIFT_NAME(fcmToken);
+@property(nonatomic, readonly, nullable) NSString *FCMToken NS_SWIFT_NAME(fcmToken)
+    DEPRECATED_MSG_ATTRIBUTE("Use registerWithCompletion: instead.");
 
 /**
  * Asynchronously gets the default FCM registration token.
@@ -320,8 +323,9 @@ NS_SWIFT_NAME(Messaging)
  * @param completion The completion handler to handle the token request.
  */
 
-- (void)tokenWithCompletion:(void (^)(NSString *_Nullable token,
-                                      NSError *_Nullable error))completion;
+- (void)tokenWithCompletion:
+    (void (^)(NSString *_Nullable token, NSError *_Nullable error))completion
+    DEPRECATED_MSG_ATTRIBUTE("Use registerWithCompletion: instead.");
 
 /**
  * Asynchronously deletes the default FCM registration token.
@@ -332,7 +336,8 @@ NS_SWIFT_NAME(Messaging)
  * @param completion The completion handler to handle the token deletion.
  */
 
-- (void)deleteTokenWithCompletion:(void (^)(NSError *_Nullable error))completion;
+- (void)deleteTokenWithCompletion:(void (^)(NSError *_Nullable error))completion
+    DEPRECATED_MSG_ATTRIBUTE("Use unregisterWithCompletion: instead.");
 
 /**
  *  Retrieves an FCM registration token for a particular Sender ID. This can be used to allow
@@ -356,7 +361,8 @@ NS_SWIFT_NAME(Messaging)
 - (void)retrieveFCMTokenForSenderID:(NSString *)senderID
                          completion:(void (^)(NSString *_Nullable FCMToken,
                                               NSError *_Nullable error))completion
-    NS_SWIFT_NAME(retrieveFCMToken(forSenderID:completion:));
+    NS_SWIFT_NAME(retrieveFCMToken(forSenderID:completion:))
+        DEPRECATED_MSG_ATTRIBUTE("Use registerWithCompletion: instead.");
 
 /**
  * Invalidates an FCM token for a particular Sender ID. That Sender ID cannot no longer send
@@ -368,7 +374,8 @@ NS_SWIFT_NAME(Messaging)
  */
 - (void)deleteFCMTokenForSenderID:(NSString *)senderID
                        completion:(void (^)(NSError *_Nullable error))completion
-    NS_SWIFT_NAME(deleteFCMToken(forSenderID:completion:));
+    NS_SWIFT_NAME(deleteFCMToken(forSenderID:completion:))
+        DEPRECATED_MSG_ATTRIBUTE("Use unregisterWithCompletion: instead.");
 
 #pragma mark - FID