+++ FirebaseMessaging/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+- [changed] Deprecate token-related APIs (`tokenWithCompletion:`, `deleteTokenWithCompletion:`, `retrieveFCMTokenForSenderID:completion:`, `deleteFCMTokenForSenderID:completion:`, `FCMToken`, `messaging:didReceiveRegistrationToken:`, and `FIRMessagingRegistrationTokenRefreshedNotification`) in favor of the new FID-based registration APIs.
 - [fixed] Fix an issue where new FCM registration processing was not invoked
   when `FirebaseMessagingInstallationIdEnabled` was set to `YES` and a legacy token
   existed in cache. (#16429)